ai-service/requirements.txt
salesforce-lavis
deepspeech
torch